+++ docs/CHANGELOG.md
@@ -4,7 +4,9 @@
   - jwt_tool
   - MANSPIDER
   - ncrack
-
+  - weevely
+- **New features**
+  - Locations now support [globbing](https://ruby-doc.org/3.3.0/Dir.html#method-c-glob), useful when needing to remove several files in a directory while not removing the whole directory (e.g. log files with rotation in the same directory as the configuration file)
 - **Quality**
   - The configuration now has a version and there are several checks to detected mismatch between the configuration version and the tool version. This prevents the user performing complex debugging after an update.